Add files via upload
In this project I have handled data manipulation and data preprocessing part where I have cleaned and removed unnecessary data from the data and found some meaningful insights from the data through Python libraries (Pandas, Numpy). In this project I check the month wisesales and the month which got the most number of orders and the month with the least orders.